requirements.txt
asgiref==3.4.1
certifi==2023.7.22
charset-normalizer==3.2.0
dict2xml==1.7.3
Django==3.2.5
future==0.18.3
idna==3.4
image==1.5.33
imageio==2.28.1
intelhex==2.3.0
iso8601==1.1.0
numpy==1.24.2
paho-mqtt==1.6.1
Pillow==9.5.0
pyserial==3.5
pytz==2021.1
PyYAML==6.0
rawpy==0.18.1
regex==2021.7.6
requests==2.31.0
serial==0.0.97
six==1.16.0
sqlparse==0.4.1
urllib3==2.0.4
XlsxWriter==3.1.0
xmltodict==0.13.0
